Player Size, Some Assets and Marching Cubes Terrain
Quite a large commit this time around, since I've done a lot. I have implemented two variants of a cactus and two variants of a tree, in addition to a WIP of the mech model, mainly serves to give a sense of scale to the other assets. I have adjusted the size of the player and the position of the camera to emulate what it feels like to be as large as the mech. The procedural terrain uses the marching cubes algorithm and is optimised with chunking and welded vertices. Has many many toggles for shading, smoothness, density, amplitude, etc... feel free to play around! Settings can be hot reloaded with R, although toggling the "Use Blur" button in runtime crashes the code, I know the fix for it though, will implement it in the next commit.

The terrain is for a prototype Tyler and I thought of in week 9, where the player would spawn into a random world, fight enemies and collect items. This mode would be chosen from a Hub world. We will use a select few seeds and design levels around those seeds, which negates the worry of trying to procedurally place objects. In essence, we are using the randomly generated terrain and building the world around its topography, which we hope gives a unique enough experience with some replayability.