Bug Fixes
Fixed a few bugs where the mesh would be generated with a bias towards the lower quarter of the chunk bounds. Also added a falloff for the edges of the generated mesh to seamlessly blend to Y=0, which is ideal if a body of water to be added to the perimeter of the procedural island.