Reduced Player Size + Changed Default Seed
Changed the default seed for the world gen to a nicer one. Also reduced the player size back to its original value, as there are problems with the ground detector script at large sizes. Will look into this.

I think the procedural terrain generation code is fine at the moment, but it could benefit from some more modularity and readability. I plan on adding an interface to help designers with creating more interesting levels, there are a lot of undocumented magic numbers in the code that I used for testing, and these need to be made more human readable.