Lots n lots of bug fixes and features
a last minute effort to get as much as the missing functionality in as possible.

This commit has been a few days in the making on a separate local copy of the project.
I have added the following:
Refined the object scatterer to have rotation, scale, limit number of objects.
A* path finding for procedural dirt roads between certain Poisson Disc scattered layers
A* path finding for enemy movement and attack logic for procedural levels
added grass to the procedural level
destroyable objects in procedural level
attack animation
booster animation
main menu